+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Change Log
 ==========
 
+Version 1.1.0 *(In Development)*
+--------------------------------
+
+ * Extend from `ViewPager` so it can be used with other APIs which require an
+   instance of `ViewPager`.
+ * Move to `com.jakewharton.android.view` package.
+
+
 Version 1.0.0 *(2011-08-24)*
 ----------------------------